renaming folders & readme update
